is include_type_species and is_include_type_genus buttons introduced. They filter out only if non-reference material is selected.
Since all reference data should fully defined and no type exists in the reference data. The logic is only called if non-reference material is searched for. There either material is returned or type material is filterd out depending on the users preference

Additonally, the genus drop down menus also have the is_type filtered out which will make it easier for users. So they wont have to go through all these type options that wont deliver anything anyway. This is also done for the alphabet search.
Small todo is to implement this for species for the dropdowns it already filters items by istype species.

Also in place is  logic for the webpage that if non-reference material is not wanted then the type options will dissapear

Also done was each drop down menu starts with an intial "please select from ..." this makes it cleaner

also randomisation fixed